requirements.txt
fastapi
uvicorn
httpx
google-genai
python-dotenv
pydantic
tenacity